feat: close problem deploy feedback loop and prepare Hono services for Lambda
Problem deploy previously fired CreateStack and returned immediately, so jobs
sat in_progress forever in EventBridge mode. This PR:

- Wait for CREATE_COMPLETE in both the TS handler (via
  waitUntilStackCreateComplete) and the CodeBuild bash script (via
  DescribeStacks polling). exit 1 on terminal failure so the ScriptJob emits
  problem.deploy.failed.
- Add POST /api/internal/deploy-events receiver that consumes
  problem.deploy.completed / failed EventBridge payloads and updates the
  DeploymentJob row in DynamoDB. Shared-secret auth via
  INTERNAL_EVENT_TOKEN header (production requires it to be set).
- Introduce a composite deploymentKey (eventId:problemId:jobId) as
  ScriptJob jobIdentifierKey so both success and failure events carry enough
  context to map back to a single job without scans.
- Drop SSE (/deployments/stream and in-memory subscriber Map) and rely on
  the existing 5-second polling fallback — needed for Lambda compatibility.
- Extract each Hono microservice app construction into app.ts, keep
  index.ts for local node-server, and add lambda.ts that exports
  handle(app) as the Lambda handler.
- Split admin-deploy.ts into AWS ops + a new admin-gameday-deploy.ts, and
  hoist deployments page helpers into a sibling helpers.ts module to keep
  files under the oversized-module threshold.

Follow-up required from infra side: EventBridge rule → API Destination →
/api/internal/deploy-events wiring, OpenNext for the Next.js apps, and
removing the leaderboard-service SSE routes.
