+++ server/application/microservices/battle-service/src/index.ts
@@ -1,31 +1,9 @@
 import { serve } from '@hono/node-server';
-import { Hono } from 'hono';
-import { cors } from 'hono/cors';
+import { app } from './app';
 import { createLogger } from './lib/logger';
-import { healthRoutes } from './api/health';
-import { battlesRoutes } from './api/battles';
-import { authMiddleware } from './middleware/auth';
 
 const logger = createLogger('battle-service');
-const app = new Hono();
 
-// CORS設定
-app.use(
-  '/*',
-  cors({
-    origin: process.env.CORS_ORIGIN?.split(',') ?? ['http://localhost:13000'],
-    credentials: true,
-  })
-);
-
-// バトルルートに認証を適用
-app.use('/battles/*', authMiddleware);
-
-// ルート登録
-app.route('/', healthRoutes);
-app.route('/', battlesRoutes);
-
-// PORT バリデーション
 const parsePort = (value: string | undefined): number => {
   const defaultPort = 3010;
   if (!value) return defaultPort;
@@ -34,7 +12,7 @@ const parsePort = (value: string | undefined): number => {
   if (Number.isNaN(parsed) || parsed < 1 || parsed > 65535) {
     logger.warn(
       { providedValue: value },
-      `無効な PORT 値です。デフォルト (${defaultPort}) を使用します`
+      `無効な PORT 値です。デフォルト (${defaultPort}) を使用します`,
     );
     return defaultPort;
   }
@@ -50,10 +28,9 @@ const server = serve(
   },
   (info) => {
     logger.info({ port: info.port }, 'バトル管理サービスが起動しました');
-  }
+  },
 );
 
-// グレースフルシャットダウン
 const shutdown = () => {
   logger.info('シャットダウンを開始します');
   server.close(() => {
