+++ server/application/microservices/scoring-service/src/index.ts
@@ -1,35 +1,9 @@
 import { serve } from '@hono/node-server';
-import { Hono } from 'hono';
-import { cors } from 'hono/cors';
+import { app } from './app';
 import { createLogger } from './lib/logger';
-import { healthRoutes } from './api/health';
-import { scoringRoutes } from './api/scoring';
-import { scoresRoutes } from './api/scores';
-import { authMiddleware } from './middleware/auth';
 
 const logger = createLogger('scoring-service');
-const app = new Hono();
 
-// CORS設定
-app.use(
-  '/*',
-  cors({
-    origin: process.env.CORS_ORIGIN?.split(',') ?? ['http://localhost:13000'],
-    credentials: true,
-  })
-);
-
-// 採点ルートに認証を適用
-app.use('/criteria/*', authMiddleware);
-app.use('/sessions/*', authMiddleware);
-app.use('/api/scores/*', authMiddleware);
-
-// ルート登録
-app.route('/', healthRoutes);
-app.route('/', scoringRoutes);
-app.route('/', scoresRoutes);
-
-// PORT バリデーション
 const parsePort = (value: string | undefined): number => {
   const defaultPort = 3011;
   if (!value) return defaultPort;
@@ -38,7 +12,7 @@ const parsePort = (value: string | undefined): number => {
   if (Number.isNaN(parsed) || parsed < 1 || parsed > 65535) {
     logger.warn(
       { providedValue: value },
-      `無効な PORT 値です。デフォルト (${defaultPort}) を使用します`
+      `無効な PORT 値です。デフォルト (${defaultPort}) を使用します`,
     );
     return defaultPort;
   }
@@ -54,10 +28,9 @@ const server = serve(
   },
   (info) => {
     logger.info({ port: info.port }, '採点サービスが起動しました');
-  }
+  },
 );
 
-// グレースフルシャットダウン
 const shutdown = () => {
   logger.info('シャットダウンを開始します');
   server.close(() => {
