+++ server/application/microservices/tenant-management/src/lambda.ts
@@ -1,0 +1,7 @@
+/**
+ * AWS Lambda handler entry.
+ */
+import { handle } from 'hono/aws-lambda';
+import { app } from './app';
+
+export const handler = handle(app);
